+++ com.unity.perception/CHANGELOG.md
@@ -21,6 +21,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Fixed
 
+Fixed the math offsetting the iteration index of each Unity Simulation instance directly after they deserialize their app-params
+
 ## [0.7.0-preview.1] - 2021-02-01
 
 ### Upgrade Notes